feat: unit tests added

+++ tests/unit/import.test.ts
@@ -1,12 +1,11 @@
 import { jest } from "@jest/globals";
 import * as factory from "../factories/import.factory.js";
-import * as importServices from "../../src/services/import.services.js"
+import { importServices } from "../../src/services/import.services.js";
+import { productRepository } from "../../src/repositories/products.repositorires.js";
 
 import fs from "fs"
 import { faker } from "@faker-js/faker";
 
-beforeAll(() => {
-})
 
 describe("Import test", () => {
 	it("Given an zipped Buffer, should unzip buffer", async () => {
@@ -24,11 +23,29 @@ describe("Import test", () => {
 		expect(fs.existsSync('temp/test_file.txt')).toEqual(true);
 	})
 
+	it("Given a file, should readLine by line",async () => {
+		jest.spyOn(productRepository, "findProductByCode").mockImplementation(async () => Promise.resolve(null));
+		const FILE_NAME = "test_file"
+		await factory.createFileAndFillWith100Products(FILE_NAME)
+		const products = await importServices.readFileAndReturn100Objects(FILE_NAME)
+		expect(products.length).toEqual(100);
+	})
+
+	it("Given a file, should delete it",async () => {
+		const FILE_NAME = "test_file"
+		await factory.createFileAndFillWith100Products(FILE_NAME)
+		await importServices.deleteFile(FILE_NAME)
+		expect(fs.existsSync('temp/test_file.txt')).toBe(false);
+	})
+
 	
 });
 
 
 
 afterAll(() => {
-	fs.promises.unlink(`temp/test_file.txt`)
+	if(fs.existsSync('temp/test_file.txt')){
+		fs.promises.unlink(`temp/test_file.txt`)
+	}
+	jest.clearAllMocks();
 })
